java: Implement pocket bridge
Now the service is just a layer between proximity sensor checks and
pocket judge kernel driver.

When IN POCKET (sensor covered), we write "1" to sysfs node to trigger
callbacks in the driver, which should disable IRQ interruptions for
desired parts of the device (touchscreen, HW buttons, fingerprints,
sliders etc.).

When NOT IN POCKET (sensor uncovered), we write "0" to sysfs node which
should do the opposite effect.

When the user presses power button for n seconds (2s by default), the
pocket judge driver will output "2" when reading sysfs node and all IRQ
interruptions are released immediately; this state is called "safe
door". In our java app we run a thread to listen for "safe door" state
if proximity sensor was positive (covered) at least ones.

When IN "SAFE DOOR" state, we write "0" to sysfs node and temporarily
suspend pocket mode and disable proximity sensor until next screen
off/on cycle.

Signed-off-by: iusmac <iusico.maxim@libero.it>