java: Use volume keys (if possible) to wake screen
Force wake screen with volume keys if volume buttons should not seek
media tracks. This will allow to trigger pocket judge and prevent from
changing volume song while in pocket.

This settings is implement in LineageOS-based ROMs only. Revert this
commit if it's missing in your ROM to allow compilation.

Signed-off-by: iusmac <iusico.maxim@libero.it>