java: Fix safe-door thread not properly closed
Currently mIsSafeDoorThreadExit variable is changed in the main thread
and checked inside a background thread, but due to the CPU's internal
cache the memory visibility may be compromised, thus it's very likely
that the thread won't terminate, which will also leak memory since we're
creating a new one every time. This also explains why the keyguard
starts to hang completely for a couple of seconds after some lock/unlock
cycles.

Signed-off-by: iusmac <iusico.maxim@libero.it>